More fixes to game balance, and no more Curse of Undeath
The Curse of Undeath is gone now; as long as capital cities are immortal, it's unfair to saddle them with an unbreakable curse, so unless that rule ever gets changed, the Curse of Undeath is not coming back. The Undead are weaker in general, with a smaller bonus against cities and a debuffing against anything besides cities - after all, everyone knows that individual zombies are weak, and it's only in numbers that they're fearsome. The debuffs on science and culture have been redone, so they no longer require percentages or specific buildings.

Also gone is the Fast Fossil Fuel, because that one is moving to its own mod. I'll be making a mod that adds that building and at least one other, all of which give you the ability to change outdated resources into more currently useful ones.